+++ src/php/Lang/TypeFactory.php
@@ -11,6 +11,8 @@ use Phel\Lang\Collections\LinkedList\PersistentListInterface;
 use Phel\Lang\Collections\Map\PersistentArrayMap;
 use Phel\Lang\Collections\Map\PersistentHashMap;
 use Phel\Lang\Collections\Map\PersistentMapInterface;
+use Phel\Lang\Collections\SortedMap\PersistentSortedMap;
+use Phel\Lang\Collections\SortedSet\TransientSortedSet;
 use Phel\Lang\Collections\Vector\PersistentVector;
 use Phel\Lang\Collections\Vector\PersistentVectorInterface;
 
@@ -74,6 +76,28 @@ final class TypeFactory
     public function persistentVectorFromArray(array $values): PersistentVectorInterface
     {
         return PersistentVector::fromArray($this->hasher, $this->equalizer, $values);
+    }
+
+    /**
+     * @param ?callable(mixed, mixed): int $comparator
+     */
+    public function persistentSortedMapFromArray(array $kvs = [], ?callable $comparator = null): PersistentMapInterface
+    {
+        return PersistentSortedMap::fromArray($this->hasher, $this->equalizer, $kvs, $comparator);
+    }
+
+    /**
+     * @param ?callable(mixed, mixed): int $comparator
+     */
+    public function persistentSortedSetFromArray(array $values = [], ?callable $comparator = null): PersistentHashSetInterface
+    {
+        $map = PersistentSortedMap::empty($this->hasher, $this->equalizer, $comparator);
+        $transient = new TransientSortedSet($this->hasher, $map->asTransient());
+        foreach ($values as $value) {
+            $transient->add($value);
+        }
+
+        return $transient->persistent();
     }
 
     public function getEqualizer(): EqualizerInterface